take advantage of c_free() and c_realloc() support of nullptr
free(NULL) is a no-op, it will not cause a segfault or any undefined behavior

Likewise, realloc(NULL, size) is treated as if it were malloc(size)

As such, we can avoid premallocing for param_ptrs, as pgdata() will call
realloc() if it fills in that param. This is extra beneficial for strings
as we avoid allocating a small string param only to turn around and realloc
it for a larger actual string.